delet e edit

+++ Topicos Especiais em desenvolvimento WEB/aula 03_05/projeto-laravel/app/Http/Controllers/CartasController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Carta;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 
 class CartasController extends Controller
 {
@@ -35,10 +36,22 @@ class CartasController extends Controller
 
     }
 
-    public function excluir($id) {
-        $carta = Carta::findOrFail($id);
-        $carta->delete();
+    public function editar(Request $request, Carta $carta){
 
-        return redirect()->route('cartas.index')->with('success', 'Carta excluída com sucesso!');
+    }
+
+    public function excluir(Request $request, Carta $carta) {
+        if ($request->isMethod('DELETE')) {
+            if ($carta->foto) {
+                Storage::disk('public')->delete($carta->foto);
+                // unlink(storage_path('app/public/' . $carta->foto));
+            }
+            $carta->delete();
+            return redirect()->route('cartas.index')->with('success', 'Carta excluída com sucesso!');
+        }
+
+        return view('cartas.excluir', [
+            'carta' => $carta
+        ]);
     }
 }
